requirements.txt
albumentations==1.4.7
comet_ml==3.42.1
coremltools==7.2
Django==3.2.23
Django==4.2.11
Flask==2.0.3
Flask==1.1.2
GitPython==3.1.43
GitPython==3.1.43
ipython==7.29.0
ipython==8.12.3
keras==2.10.0
keras==3.3.2
librosa==0.10.2
matplotlib==3.4.3
matplotlib==3.8.4
mss==9.0.1
nncf==2.10.0
numpy==1.20.3
onnx==1.16.1
onnxruntime==1.18.0
onnxsim==0.4.36
opencv_python==4.7.0.68
opencv_python==4.9.0.80
openvino==2024.1.0
paddle==1.0.2
pafy==0.5.5
pandas==1.3.4
Pillow==9.4.0
Pillow==8.4.0
Pillow==10.3.0
psutil==5.8.0
psutil==5.9.8
pycocotools==2.0.7
pytesseract==0.3.10
PyYAML==6.0
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.32.2
scikit_learn==0.24.2
scikit_learn==1.4.2
scipy==1.7.1
seaborn==0.11.2
setuptools==68.0.0
setuptools==58.0.4
setuptools==61.2.0
setuptools==65.6.3
setuptools==69.5.1
six==1.15.0
six==1.16.0
SpeechRecognition==3.10.3
tensorflow==2.10.1
tensorflow==2.16.1
tensorflow_intel==2.16.1
tensorflowjs==4.19.0
tensorrt==10.0.1
tflite_runtime==2.14.0
tflite_support==0.4.4
thop==0.1.1.post2209072238
torch==1.11.0
torch==2.0.1
torch==1.12.0
torch==2.3.0
torchvision==0.15.2
torchvision==0.13.0
torchvision==0.18.0
tqdm==4.62.3
tqdm==4.66.4
transformers==4.40.2
tritonclient==2.46.0
ultralytics==8.2.8
x2paddle==1.4.1
